+++ src/main/java/com/hyeon/guardrail/user/domain/UserRole.java
@@ -1,8 +1,0 @@
-package com.hyeon.guardrail.user.domain;
-
-/** 사용자 역할 */
-public enum UserRole {
-  STAFF,
-  OPERATOR,
-  ADMIN
-}