+++ src/main/java/com/hyeon/guardrail/user/domain/UserStatus.java
@@ -1,7 +1,0 @@
-package com.hyeon.guardrail.user.domain;
-
-/** 사용자 사용 상태 */
-public enum UserStatus {
-  ACTIVE,
-  INACTIVE
-}